requirements.txt
tensorflow
numpy
pandas
matplotlib
seaborn
tqdm
streamlit
pillow
